src: Implement a `-N` flag to ignore user configuration
Users with a heavy configuration that customise a lot of default
settings/side-effects cannot easily have it ignored by the editor.

This commit implements a `-N` flag that instructs the editor to load
system files, but skip those written by the user. In terms of built-in
options:

* load files in %val{runtime}
* ignore files in %val{config}

Note that, despite being user-created, the `kakrc.local` file will
still be loaded.